Bug 19893 - Alternative optimized indexing for Elasticsearch
Add alternative optimized indexing for Elasticsearch

How to test:
1) Time a full elasticsearch re-index by running the rebuild_elastic_search.pl
   with the -d flag: `koha-shell <instance_name> -c "time rebuild_elastic_search.pl -d"`.
2) Enable ExperimentalElasticSearchIndexing system preference
   (found under Global System preferences -> Administration -> Search Engine).
3) Time a full re-index again, it should be about twice at fast (for a
   couple of thousand biblios, with fewer biblios results may be more
   unpredictable).

Sponsored-by: Gothenburg University Library

+++ misc/search_tools/rebuild_elastic_search.pl
@@ -160,11 +160,14 @@ sub do_reindex {
     my ( $next, $index_name ) = @_;
 
     my $indexer = Koha::SearchEngine::Elasticsearch::Indexer->new( { index => $index_name } );
-    if ($delete) {
 
-        # We know it's safe to not recreate the indexer because update_index
-        # hasn't been called yet.
-        $indexer->drop_index();
+    if ($delete) {
+        $indexer->drop_index() if $indexer->index_exists();
+        $indexer->create_index();
+    }
+    elsif (!$indexer->index_exists()) {
+        # Create index if does not exist
+        $indexer->create_index();
     }
 
     my $count        = 0;
